DataProcessing/requirements.txt
pinecone
openai
langchain
langchain-openai
pypdf
beautifulsoup4
python-dotenv
tqdm